+++ CHANGELOG.md
@@ -17,6 +17,7 @@ We use the following categories for changes:
 ## [Unreleased]
 
 ### Added
+- Added big cluster preset [#sc-8158]
 
 ### Changed